// src/core/infrastructure/supabase/types.ts
export type Database = {
// ...
  public: {
    Tables: {
// ...
      feedbacks: {
        Row: {
          created_at: string
          customer_name: string
          customer_username: string | null
          display_order: number | null
          feedback_text: string
          id: string
          image_url: string
          is_active: boolean | null
          updated_at: string
        }
        Insert: {
          created_at?: string
          customer_name: string
          customer_username?: string | null
          display_order?: number | null
          feedback_text: string
          id?: string
          image_url: string
          is_active?: boolean | null
          updated_at?: string
        }
        Update: {
          created_at?: string
          customer_name?: string
          customer_username?: string | null
          display_order?: number | null
          feedback_text?: string
          id?: string
          image_url?: string
          is_active?: boolean | null
          updated_at?: string
        }
        Relationships: []
      }
// ...
    }
// ...
  }
}